Move SokeSelect to frontend-common
