+++ frontend/frontend-common/index.ts
@@ -1,3 +1,5 @@
 import useDebounce from "./hooks/useDebounce";
+import { SokeSelect } from "./components/SokeSelect";
+import { shallowEquals } from "./utils/shallow-equals";
 
-export { useDebounce };
+export { useDebounce, SokeSelect, shallowEquals };
