+++ frontend/frontend-common/utils/shallow-equals.ts
