--- frontend/mr-admin-flate/src/components/skjema/SokeSelect.module.scss
@@ -1,19 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-}
-
-.description {
-  margin-bottom: 8px;
-  display: inline-block;
-  color: var(--ac-form-description, var(--a-text-subtle));
-}
-
-.label {
-  display: inline-block;
-}
-
-.errormsg {
-  margin-top: 8px;
-  color: #c30000;
-}
